vmac: port VmacAccel into HwComponent form (Phase 1)
Mirror examples/shared_mem/HistAccel: VmacAccel is now a synthesizable
HwComponent (cpp_kernel_name="vmac", cpp_namespace="vmac_impl") with the
structural widths as HwParam[int] (mem_dwidth/mem_awidth/data_bw/acc_bw/
out_bw), an m_axi MMIFMaster data port + s_in command stream, run_proc as
the codegen root, and the single @synthesizable(impl_file=
"vmac_compute_impl.tpp") vmac_compute hook whose Python body is the golden
(delegates to execute). accumulator_format / output_format / the
self-checking accumulator invariant are preserved unchanged.

golden.py is now a thin re-export shim (retired in Phase 3) so the golden /
numeric / conformance imports keep working. Constructed without a sim,
VmacAccel stays a lightweight params + golden object (the form the tests
use); with a sim it is the full HwComponent.

Gate: vmac golden/numeric/cmd tests pass; full -m "not vitis" = only the 15
baseline failures.
